Most basic stuff working
About to start diff

+++ src/Kanoop/network.cpp
@@ -12,7 +12,7 @@ Network::Network(Repository* repo) :
     _remotes = new RemoteCollection(repo);
     git_strarray list;
     if(git_remote_list(&list, repo->handle().value()) == 0) {
-        for(int i = 0;i < list.count;i++) {
+        for(int i = 0;i < (int)list.count;i++) {
             Remote* remote = new Remote(repo, list.strings[i]);
             _remotes->append(remote);
         }
